+++ _test_common/lib/assets.dart
@@ -13,7 +13,7 @@ AssetNode makeAssetNode([
   Digest? lastKnownDigest,
 ]) {
   var id = makeAssetId(assetIdString);
-  var node = SourceAssetNode(id, lastKnownDigest: lastKnownDigest);
+  var node = AssetNode.source(id, lastKnownDigest: lastKnownDigest);
   if (outputs != null) {
     node.outputs.addAll(outputs);
     node.primaryOutputs.addAll(outputs);